information/dashboard/requirements.txt
Flask==2.3.2
pandas==2.0.3
PyYAML==6.0.1